pyproject.toml
[project]
name = "cad-image-cropper"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.0.0",
    "opencv-python-headless>=4.10.0",
    "transformers>=4.40.0",
    "torch>=2.4.0",
    "numpy>=2.1.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
cad-crop = "cad_image_cropper.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cad_image_cropper"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
files = ["src"]
